Refactoring ConversationUpdateRouteBuilder and EventRouteBuilder and updating Create() definitions
