+++ src/libraries/Builder/Microsoft.Agents.Builder/App/EventRouteBuilder.cs
@@ -2,187 +2,40 @@
 // Licensed under the MIT License.
 
 
-using Microsoft.Agents.Builder.Errors;
 using Microsoft.Agents.Core;
-using Microsoft.Agents.Core.Models;
-using System;
-using System.Text.RegularExpressions;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Microsoft.Agents.Builder.App
 {
     /// <summary>
-    /// RouteBuilder for routing Event activities in an AgentApplication.
+    /// Provides a concrete builder for routing event activities in an <see cref="AgentApplication"/>.
     /// </summary>
     /// <remarks>
-    /// Use <see cref="EventRouteBuilder"/> to create and configure routes that respond to event
-    /// activities. This builder allows matching event activities by name or regular expression, and supports
-    /// channelId and agentic routing scenarios. Instances are created via the <see cref="Create"/> method
-    /// and optionally configured using <see cref="WithName(string)"/>, <see cref="WithName(Regex)"/>,
-    /// or <see cref="WithSelector(RouteSelector)"/>. If neither <see cref="WithName(string)"/> nor
-    /// <see cref="WithName(Regex)"/> is called, the route will match any event activity regardless of name.<br/><br/>
-    /// Example usage:<br/><br/>
-    /// <code>
-    /// var route = EventRouteBuilder.Create()
-    ///    .WithName("myEvent")
-    ///    .WithHandler((context, state, ct) => context.SendActivityAsync("Event received!", cancellationToken: ct))
-    ///    .Build();
-    ///
-    /// app.AddRoute(route);
-    /// </code>
+    /// Use <see cref="EventRouteBuilder"/> when you need an event route that uses the standard
+    /// <see cref="RouteHandler"/> delegate. This type inherits the shared event matching behavior from
+    /// <see cref="EventRouteBuilderBase{TBuilder}"/>, including event-name filters, custom selectors, channel
+    /// constraints, and agentic routing support.
     /// </remarks>
-    public class EventRouteBuilder : RouteBuilderBase<EventRouteBuilder>
+    public class EventRouteBuilder : EventRouteBuilderBase<EventRouteBuilder>
     {
-        private string _eventName;
-        private Regex _eventRegex;
-
         /// <summary>
-        /// Configures the route to match event activities with the specified name, using a case-insensitive comparison.
+        /// Creates a new instance of the <see cref="EventRouteBuilder"/> class.
         /// </summary>
-        /// <remarks>This method restricts the route to only handle event activities whose name matches
-        /// the specified value. If the route is marked as agentic, only agentic requests will be considered for
-        /// matching.</remarks>
-        /// <param name="name">The name of the event activity to match. Comparison is case-insensitive. Cannot be null.</param>
-        /// <returns>A EventRouteBuilder instance with the added selector for matching Activity.Name.</returns>
-        public EventRouteBuilder WithName(string name)
+        /// <returns>A new <see cref="EventRouteBuilder"/>.</returns>
+        public static EventRouteBuilder Create()
         {
-            AssertionHelpers.ThrowIfNullOrWhiteSpace(name, nameof(name));
-
-            if (_eventName != null)
-            {
-                throw Core.Errors.ExceptionHelper.GenerateException<InvalidOperationException>(ErrorHelper.RouteSelectorAlreadyDefined, null, $"EventRouteBuilder.WithName({name}) with Name already set");
-            }
-
-            if (_eventRegex != null)
-            {
-                throw Core.Errors.ExceptionHelper.GenerateException<InvalidOperationException>(ErrorHelper.RouteSelectorAlreadyDefined, null, $"EventRouteBuilder.WithName({name}) with Name Regex already set");
-            }
-
-            _eventName = name;
-            return this;
-        }
-
-        /// <summary>
-        /// Configures the event route to match event activities whose name satisfies the specified regular expression
-        /// pattern.
-        /// </summary>
-        /// <remarks>This method restricts the route to event activities whose name matches the provided
-        /// pattern. If the route is marked as agentic, only agentic requests will be considered for matching.</remarks>
-        /// <param name="namePattern">The regular expression used to match the name of incoming event activities. Cannot be null.</param>
-        /// <returns>A EventRouteBuilder instance configured with the specified Activity.Name pattern selector.</returns>
-        public EventRouteBuilder WithName(Regex namePattern)
-        {
-            AssertionHelpers.ThrowIfNull(namePattern, nameof(namePattern));
-
-            if (_eventRegex != null)
-            {
-                throw Core.Errors.ExceptionHelper.GenerateException<InvalidOperationException>(ErrorHelper.RouteSelectorAlreadyDefined, null, $"EventRouteBuilder.WithName(Regex({namePattern})) with Name Regex already set");
-            }
-
-            if (_eventName != null)
-            {
-                throw Core.Errors.ExceptionHelper.GenerateException<InvalidOperationException>(ErrorHelper.RouteSelectorAlreadyDefined, null, $"EventRouteBuilder.WithName(Regex({namePattern})) with Name already set");
-            }
-
-            _eventRegex = namePattern;
-            return this;
-        }
-
-        /// <summary>
-        /// Sets a custom route selector used to determine how incoming requests are matched to this route builder.
-        /// </summary>
-        /// <remarks>Use this method to customize the matching logic for routes. This allows for advanced
-        /// routing scenarios where requests are selected based on custom rules or patterns. If WithName was
-        /// also called, this selector is in addition to the Name selector.</remarks>
-        /// <param name="selector">The route selector that defines the criteria for matching requests to the route. The supplied selector does
-        /// not need to validate base route properties like ChannelId, Agentic, etc. An Activity type of "event" is enforced.</param>
-        /// <returns>A EventRouteBuilder instance configured with the custom selector.</returns>
-        public override EventRouteBuilder WithSelector(RouteSelector selector)
-        {
-            AssertionHelpers.ThrowIfNull(selector, nameof(selector));
-
-            if (_route.Selector != null)
-            {
-                throw Core.Errors.ExceptionHelper.GenerateException<InvalidOperationException>(ErrorHelper.RouteSelectorAlreadyDefined, null, $"EventRouteBuilder.WithSelector()");
-            }
-
-            async Task<bool> ensureEvent(ITurnContext context, CancellationToken cancellationToken)
-            {
-                return IsContextMatch(context, _route) && context.Activity.IsType(ActivityTypes.Event) && await selector(context, cancellationToken).ConfigureAwait(false);
-            }
-
-            _route.Selector = ensureEvent;
-            return this;
+            return new EventRouteBuilder();
         }
 
         /// <summary>
         /// Assigns the specified route handler to the current route and returns the updated builder instance.
         /// </summary>
         /// <param name="handler">The route handler to associate with the route. Cannot be null.</param>
-        /// <returns>The current RouteBuilder instance with the handler set, enabling method chaining.</returns>
+        /// <returns>The current <see cref="EventRouteBuilder"/> instance for method chaining.</returns>
         public EventRouteBuilder WithHandler(RouteHandler handler)
         {
             AssertionHelpers.ThrowIfNull(handler, nameof(handler));
             _route.Handler = handler;
             return this;
-        }
-
-        /// <summary>
-        /// For event routes, the invoke flag is ignored to prevent misconfiguration.
-        /// </summary>
-        /// <remarks>Events cannot be configured as invoke routes. This method always returns the
-        /// current instance, regardless of the value of <paramref name="isInvoke"/>.</remarks>
-        /// <param name="isInvoke">Ignored</param>
-        /// <returns>The current instance of <see cref="Microsoft.Agents.Builder.App.EventRouteBuilder"/>.</returns>
-        public override EventRouteBuilder AsInvoke(bool isInvoke = true)
-        {
-            return this;
-        }
-
-        protected override void PreBuild()
-        {
-            // When no name filter is specified the route matches any event — default to Last so
-            // specific-name routes take priority without callers having to set the rank explicitly.
-            if (_eventName == null && _eventRegex == null && _route.Rank == RouteRank.Unspecified)
-            {
-                _route.Rank = RouteRank.Last;
-            }
-
-            if (_route.Selector != null)
-            {
-                if (_eventName != null || _eventRegex != null)
-                {
-                    // Match on both the existing selector and the Activity.Name
-                    var existingSelector = _route.Selector;
-                    _route.Selector = async (context, ct) =>
-                        IsContextMatch(context, _route)
-                        && context.Activity.IsType(ActivityTypes.Event)
-                        && (_eventName != null ? _eventName.Equals(context.Activity.Name, StringComparison.OrdinalIgnoreCase) : context.Activity.Name != null && _eventRegex.IsMatch(context.Activity.Name))
-                        && await existingSelector(context, ct);
-                }
-                return;
-            }
-
-            if (_eventName == null && _eventRegex == null)
-            {
-                // If no name or regex specified, just match on any event
-                _route.Selector = (context, ct) => Task.FromResult
-                    (
-                        IsContextMatch(context, _route)
-                        && context.Activity.IsType(ActivityTypes.Event)
-                    );
-                return;
-            }
-
-            // Just match on Activity.Name value
-            _route.Selector = (context, ct) => Task.FromResult
-                (
-                    IsContextMatch(context, _route)
-                    && context.Activity.IsType(ActivityTypes.Event)
-                    && context.Activity.Name != null
-                    && (_eventName != null ? _eventName.Equals(context.Activity.Name, StringComparison.OrdinalIgnoreCase) : context.Activity.Name != null && _eventRegex.IsMatch(context.Activity.Name))
-                );
         }
     }
 }
